api/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
elasticsearch>=8.13.0
sentence-transformers>=2.7.0
torch>=2.2.0
pydantic>=2.7.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
PyPDF2>=3.0.0
google-api-python-client>=2.128.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
itsdangerous>=2.1.2